Fix: Command order in test_command package
