+++ tests/test_commandline/test_mnv_cli.py
@@ -26,24 +26,14 @@ def load_and_sort_json(file_path, key):
 # 定義配置檔案與對應鍵
 CONFIG_FILES = {
     "version": ("config/test_mnv_cli_version.json", "ID"),
-    "reset_device": ("config/test_mnv_cli_reset_device.json", "ID"),
-    "reset_pcie": ("config/test_mnv_cli_reset_pcie.json", "ID"),
-    "reset_power": ("config/test_mnv_cli_reset_power.json", "ID"),
     "file_paths": ("config/test_compare_file.json", None),
     "info": ("config/test_mnv_cli_info.json", "ID"),
     "identify": ("config/test_mnv_cli_identify.json", "ID"),
     "adapter": ("config/test_mnv_cli_adapter.json", "ID"),
     "log": ("config/test_mnv_cli_log.json", "ID"),
-    "bga_on": ("config/test_mnv_cli_bga_on.json", "ID"),
-    "bga_off": ("config/test_mnv_cli_bga_off.json", "ID"),
-    "bga_high": ("config/test_mnv_cli_bga_high.json", "ID"),
-    "bga_low": ("config/test_mnv_cli_bga_low.json", "ID"),
-    "bga_medium": ("config/test_mnv_cli_bga_medium.json", "ID"),
-    "bga_invalid": ("config/test_mnv_cli_bga_invalid.json", "ID"),
     "event": ("config/test_mnv_cli_event.json", "ID"),
     "debug_error": ("config/test_mnv_cli_debug_error.json", "ID"),
     "debug_normal": ("config/test_mnv_cli_debug_normal.json", "ID"),
-    "smart_invalid": ("config/test_mnv_cli_smart_invalid.json", "ID"),
     "oem_data": ("config/test_mnv_cli_oem_data.json", "ID"),
     "led": ("config/test_mnv_cli_led.json", "ID"),
     "passthru": ("config/test_mnv_cli_passthru.json", "ID"),
@@ -117,147 +107,6 @@ class TestCLIVersion:
 
 
 @pytest.mark.order(3)
-class TestCLIResetDevice:
-    '''docstring'''
-    @pytest.mark.parametrize('test_case', SORTED_DATA["reset_device"])
-    def test_commandline(self, mnv_cli, test_case):
-        '''docstring'''
-        reset_device_result = mnv_cli.interpret(test_case["Command"])
-        logger.debug('reset_device_result = %s', reset_device_result)
-        assert reset_device_result == test_case["Expected"]
-
-
-@pytest.mark.skip(reason="Deprecated")
-@pytest.mark.order(4)
-class TestCLIResetPCIe:
-    '''docstring'''
-    @pytest.mark.parametrize('test_case', SORTED_DATA["reset_pcie"])
-    def test_commandline(self, mnv_cli, test_case):
-        '''docstring'''
-        reset_pcie_result = mnv_cli.interpret(test_case["Command"])
-        logger.debug('reset_pcie_result = %s', reset_pcie_result)
-        assert reset_pcie_result == test_case["Expected"]
-
-
-@pytest.mark.order(5)
-class TestCLIResetPower:
-    '''docstring'''
-    @pytest.mark.parametrize('test_case', SORTED_DATA["reset_power"])
-    def test_commandline(self, mnv_cli, test_case):
-        '''docstring'''
-        reset_power_result = mnv_cli.interpret(test_case["Command"])
-        logger.debug('reset_power_result = %s', reset_power_result)
-        assert reset_power_result == test_case["Expected"]
-
-
-@pytest.mark.order(6)
-class TestCLISMART:
-    '''Docstring'''
-    @pytest.mark.skip(reason="Deprecated")
-    def test_get_controller_smart_info(self, mnv_cli):
-        '''docstring'''
-        smart_info = mnv_cli.get_controller_smart_info()
-        for key, value in smart_info.__dict__.items():
-            logger.info("%s = %s", key, value)
-
-        limits = {
-            "critical_warning": lambda x: 0x00 <= int(x, 16) <= 0x05,
-            "composite_temp": lambda x: int(x) < 70,
-            "available_spare": lambda x: int(x) > 10,
-            "available_spare_threshold": lambda x: int(x) > 0,  # 可以根據需求調整
-            "percentage_used": lambda x: int(x) < 100,
-        }
-
-        smart_info_fields = smart_info.__dict__
-        for field, value in smart_info_fields.items():
-            assert limits[field](value), (f"{field.replace('_', ' ').title()} "
-                                          f"{value} is out of range!")
-
-    @pytest.mark.skip(reason="Deprecated")
-    def test_get_backend_smart_info(self, mnv_cli):
-        '''docstring'''
-        smart_info = mnv_cli.get_backend_smart_info(pd_id='1')
-        for key, value in smart_info.__dict__.items():
-            logger.info("%s = %s", key, value)
-
-        limits = {
-            "critical_warning": lambda x: 0x00 <= int(x, 16) <= 0x05,
-            "composite_temp": lambda x: int(x) < 70,
-            "available_spare": lambda x: int(x) > 10,
-            "available_spare_threshold": lambda x: int(x) > 0,  # 可以根據需求調整
-            "percentage_used": lambda x: int(x) < 100,
-            "data_units_read": lambda x: int(x) >= 0,  # 讀取數據單元必須為非負數
-            "data_units_written": lambda x: int(x) >= 0,  # 寫入數據單元必須為非負數
-            "host_read_commands": lambda x: int(x) >= 0,  # 主機讀命令必須為非負數
-            "host_write_commands": lambda x: int(x) >= 0,  # 主機寫命令必須為非負數
-            "controller_busy_time": lambda x: int(x) >= 0,  # 控制器忙碌時間必須為非負數
-            "power_cycles": lambda x: int(x) >= 0,  # 上電循環次數必須為非負數
-            "power_on_hours": lambda x: int(x) >= 0,  # 開機時間必須為非負數
-            "unsafe_shutdowns": lambda x: int(x) >= 0,  # 不安全關機次數必須為非負數
-            "media_and_data_integrity_errors": lambda x: int(x) == 0,
-            "num_err_log_entries": lambda x: int(x) == 0,  # 錯誤記錄條目應為 0
-            "warning_composite_temp_time": lambda x: int(x) >= 0,
-            "critical_composite_temp_time": lambda x: int(x) >= 0,
-            "tmp_1_transition_count": lambda x: int(x) >= 0,  # 溫度管理區域 1 次數
-            "tmp_2_transition_count": lambda x: int(x) >= 0,  # 溫度管理區域 2 次數
-            "total_time_for_tmp1": lambda x: int(x) >= 0,  # 溫度管理區域 1 總時間
-            "total_time_for_tmp2": lambda x: int(x) >= 0,  # 溫度管理區域 2 總時間
-        }
-
-        smart_info_fields = smart_info.__dict__
-
-        # 逐項檢查是否符合 limits 規範
-        for field, value in smart_info_fields.items():
-            assert limits[field](value), (f"{field.replace('_', ' ').title()} "
-                                          f"{value} is out of range!")
-
-    def test_get_backend_smart_info1(self, mnv_cli):
-        '''docstring'''
-        smart_info = mnv_cli.get_backend_smart_info(pd_id='1')
-        for key, value in smart_info.__dict__.items():
-            logger.info("%s = %s", key, value)
-
-        limits = mnv_cli.import_limits(
-            'config/test_smart_limits.json')
-
-        for field, value in smart_info.__dict__.items():
-            assert limits[field](value), (f"{field.replace('_', ' ').title()}"
-                                          f" {value} is out of range!")
-
-    def test_get_controller_smart_info1(self, mnv_cli):
-        '''docstring'''
-        smart_info = mnv_cli.get_controller_smart_info()
-        for key, value in smart_info.__dict__.items():
-            logger.info("%s = %s", key, value)
-
-        limits = mnv_cli.import_limits(
-            'config/test_smart_limits.json')
-
-        for field, value in smart_info.__dict__.items():
-            assert limits[field](value), (f"{field.replace('_', ' ').title()}"
-                                          f" {value} is out of range!")
-
-    @pytest.mark.skip(reason="Only for case generation")
-    def test_export_smart_limits(self, mnv_cli):
-        '''docstring'''
-        cmd_output = mnv_cli.export_smart_limits(
-            'config/test_smart_limits.json')
-        print(cmd_output)
-
-    def test_import_limits(self, mnv_cli):
-        '''docstring'''
-        limits = mnv_cli.import_limits(
-            'config/test_smart_limits.json')
-        for key, value in limits.items():
-            logger.info("%s = %s", key, value)
-
-        assert len(limits) == 21, 'The number of limits is not correct!'
-
-        for key, value in limits.items():
-            assert callable(value), f'{key} is not a function!'
-
-
-@pytest.mark.order(7)
 @pytest.mark.dependency(name="dump data")
 class TestCLIDumpHBA:
     '''docstring'''
@@ -269,7 +118,7 @@ class TestCLIDumpHBA:
         assert dump_hba_result == test_case["Expected"]
 
 
-@pytest.mark.order(8)
+@pytest.mark.order(4)
 @pytest.mark.dependency(depends=["dump data"])
 class TestCLIExport:
     '''docstring'''
@@ -281,7 +130,7 @@ class TestCLIExport:
         assert result, 'The two files are not the same!'
 
 
-@pytest.mark.order(9)
+@pytest.mark.order(5)
 class TestCLILog:
     '''docstring'''
     @pytest.mark.parametrize('test_case', SORTED_DATA["log"])
@@ -292,73 +141,7 @@ class TestCLILog:
         assert log_result == test_case["Expected"]
 
 
-@pytest.mark.order(10)
-class TestCLIBGAOff:
-    '''docstring'''
-    @pytest.mark.parametrize('test_case', SORTED_DATA["bga_off"])
-    def test_commandline(self, mnv_cli, test_case):
-        '''docstring'''
-        bga_off_result = mnv_cli.interpret(test_case["Command"])
-        logger.debug('bga_off_result = %s', bga_off_result)
-        assert bga_off_result == test_case["Expected"]
-
-
-@pytest.mark.order(11)
-class TestCLIBGAOn:
-    '''docstring'''
-    @pytest.mark.parametrize('test_case', SORTED_DATA["bga_on"])
-    def test_commandline(self, mnv_cli, test_case):
-        '''docstring'''
-        bga_on_result = mnv_cli.interpret(test_case["Command"])
-        logger.debug('bga_on_result = %s', bga_on_result)
-        assert bga_on_result == test_case["Expected"]
-
-
-@pytest.mark.order(12)
-class TestCLIBGAHigh:
-    '''docstring'''
-    @pytest.mark.parametrize('test_case', SORTED_DATA["bga_high"])
-    def test_commandline(self, mnv_cli, test_case):
-        '''docstring'''
-        bga_high_result = mnv_cli.interpret(test_case["Command"])
-        logger.debug('bga_high_result = %s', bga_high_result)
-        assert bga_high_result == test_case["Expected"]
-
-
-@pytest.mark.order(13)
-class TestCLIBGALow:
-    '''docstring'''
-    @pytest.mark.parametrize('test_case', SORTED_DATA["bga_low"])
-    def test_commandline(self, mnv_cli, test_case):
-        '''docstring'''
-        bga_low_result = mnv_cli.interpret(test_case["Command"])
-        logger.debug('bga_low_result = %s', bga_low_result)
-        assert bga_low_result == test_case["Expected"]
-
-
-@pytest.mark.order(14)
-class TestCLIBGAMedium:
-    '''docstring'''
-    @pytest.mark.parametrize('test_case', SORTED_DATA["bga_medium"])
-    def test_commandline(self, mnv_cli, test_case):
-        '''docstring'''
-        bga_medium_result = mnv_cli.interpret(test_case["Command"])
-        logger.debug('bga_medium_result = %s', bga_medium_result)
-        assert bga_medium_result == test_case["Expected"]
-
-
-@pytest.mark.order(15)
-class TestCLIBGAInvalid:
-    '''docstring'''
-    @pytest.mark.parametrize('test_case', SORTED_DATA["bga_invalid"])
-    def test_commandline(self, mnv_cli, test_case):
-        '''docstring'''
-        bga_invalid_result = mnv_cli.interpret(test_case["Command"])
-        logger.debug('bga_invalid_result = %s', bga_invalid_result)
-        assert bga_invalid_result == test_case["Expected"]
-
-
-@pytest.mark.order(16)
+@pytest.mark.order(6)
 class TestCLIEvent:
     '''docstring'''
     @pytest.mark.parametrize('test_case', SORTED_DATA["event"])
@@ -369,7 +152,7 @@ class TestCLIEvent:
         assert event_result == test_case["Expected"]
 
 
-@pytest.mark.order(17)
+@pytest.mark.order(7)
 class TestCLIDebugError:
     '''docstring'''
     @pytest.mark.parametrize('test_case', SORTED_DATA["debug_error"])
@@ -380,7 +163,7 @@ class TestCLIDebugError:
         assert debug_error_result == test_case["Expected"]
 
 
-@pytest.mark.order(18)
+@pytest.mark.order(8)
 class TestCLIDebugNormal:
     '''docstring'''
     @pytest.mark.parametrize('test_case', SORTED_DATA["debug_normal"])
@@ -391,18 +174,7 @@ class TestCLIDebugNormal:
         assert debug_normal_result == test_case["Expected"]
 
 
-@pytest.mark.order(19)
-class TestCLISMARTInvalid:
-    '''docstring'''
-    @pytest.mark.parametrize('test_case', SORTED_DATA["smart_invalid"])
-    def test_commandline(self, mnv_cli, test_case):
-        '''docstring'''
-        smart_invalid_result = mnv_cli.interpret(test_case["Command"])
-        logger.debug('smart_invalid_result = %s', smart_invalid_result)
-        assert smart_invalid_result == test_case["Expected"]
-
-
-@pytest.mark.order(20)
+@pytest.mark.order(9)
 class TestCLILED:
     '''docstring'''
     @pytest.mark.parametrize('test_case', SORTED_DATA["led"])
@@ -413,7 +185,7 @@ class TestCLILED:
         assert led_result == test_case["Expected"]
 
 
-@pytest.mark.order(21)
+@pytest.mark.order(10)
 class TestCLIPassthru:
     '''docstring'''
     @pytest.mark.parametrize('test_case', SORTED_DATA["passthru"])
